requirements.txt
kfp==1.8.12
google-cloud-pipeline-components==1.0.2
google-cloud-aiplatform==1.11.0
